-listClassroom を Multiprocessing により高速化した(現時点では maxProcess = 10 としている)

+++ createCourse.py
@@ -9,15 +9,17 @@ from googleapiclient.discovery import build
 from googleapiclient.errors import HttpError
 from google_auth_oauthlib.flow import InstalledAppFlow
 from google.auth.transport.requests import Request
+#
 from docopt import docopt
 from tqdm import tqdm
+from multiprocessing import Pool
 
 __doc__ = """{f}
 
 Usage:
-    {f} all [--dry-run]
+    {f} all [--dry-run] [--foreign-domain]
     {f} create [<class_file>] [--dry-run]
-    {f} enroll [<enroll_file>] [--dry-run]
+    {f} enroll [<enroll_file>] [--dry-run] [--teacher] [--foreign-domain]
     {f} remove <courses>... [--dry-run]
     {f} lists <output_csv>
     {f} info <course_id>
@@ -27,6 +29,8 @@ Options:
     all         create new courses and enroll students on the Google Classroom.
     create      create only new courses (default: classes.csv).
     enroll      enroll students on courses (default: enrollments.csv).
+                --teacher: invite / enroll Teacher(default Student)
+                --foreign-domain: force invite mode
     remove      remove courses from classroom(courseId1 courseId2 ... ).
     lists       lists of all courses.
     info        information of course information.
@@ -37,10 +41,10 @@ Options:
 
 def parseOptions():
     args = docopt(__doc__)
-#    print("  {0:<20}{1:<20}{2:<20}".format("key", "value", "type"))
-#    print("  {0:-<60}".format(""))
-#    for k,v in args.items():
-#       print("  {0:<20}{1:<20}{2:<20}".format(str(k), str(v), str(type(v))))
+    # print("  {0:<20}{1:<20}{2:<20}".format("key", "value", "type"))
+    # print("  {0:-<60}".format(""))
+    # for k, v in args.items():
+    #    print("  {0:<20}{1:<20}{2:<20}".format(str(k), str(v), str(type(v))))
     options = {}
     if args['create']:
         execMode = 'create'
@@ -48,11 +52,15 @@ def parseOptions():
             options['classFile'] = args['<class_file>']
     elif args['enroll']:
         execMode = 'enroll'
+        if args['--teacher']:
+            options['teacherRole'] = True
+        if args['--foreign-domain']:
+            options['foreignDomain'] = True
         if args['<enroll_file>']:
             options['enrollFile'] = args['<enroll_file>']
     elif args['remove']:
         execMode = 'remove'
-        options['courses'] = args['courses']
+        options['courses'] = args['<courses>']
     elif args['lists']:
         execMode = 'lists'
         options['output_csv'] = args['<output_csv>']
@@ -128,7 +136,7 @@ def api_init_enroll():
 def readData():
     global classSubjects, classSections, classTeachers
     global studentEmails
-    global enrollStudents
+    global enrollUsers
     global courseLists
 
     # set filename configured by the execute option
@@ -158,11 +166,11 @@ def readData():
     # csv format:
     # classCode(Multiple Key), student id
     with open(enrollFile, 'r') as f:
-        enrollStudents = {}
+        enrollUsers = {}
         for line in f:
             line = line.rstrip('\n').split(',')
             # multiple values for single key
-            enrollStudents.setdefault(line[0], []).append(line[1])
+            enrollUsers.setdefault(line[0], []).append(line[1])
     # read already created course ID
     # csv format:
     # classCode, Google Classroom course id
@@ -224,55 +232,50 @@ def deleteClassroom(courseId):
     # serviceClassroom.courses().get(id=courseId).execute()
 
 
-def enrollStudent(courseId, studentId):
-    student = {
-        'courseId': courseId,
-        'userId': studentId,
-        'role': 'STUDENT'
+
+def inviteUser(courseId, userId, Role):
+    user = {
+        'userId': userId,
+        'role': Role
     }
     try:
-        student = serviceEnrollment.invitations().create(
-            body=student).execute()
+        user = serviceEnrollment.invitations().create(
+            body=user).execute()
         print('done')
     except HttpError as e:
         error = simplejson.loads(e.content).get('error')
         if(error.get('code') == 409):
             print('User {0} are already a member of this course.'.format(
-                student['userId']))
+                user['userId']))
         else:
             raise
 
+def enrollTeacer(courseId, teacherId):
+    teacher = {
+        'userId': teacherId
+    }
+    teacher = serviceEnrollment.courses().teachers().create(
+        courseId=courseId, body=teacher).execute()
+    print('Course {0} enroll {1} User'.format(courseId, teacherId))
 
-def enrollStudent3(courseId, studentIds):
-    student = []
-    for studentId in studentIds:
-        student.append({
-            'courseId': courseId,
-            'userId': studentId,
-            'role': 'STUDENT'
-        })
-    print(student)
-    try:
-        student = serviceEnrollment.invitations().create(
-            body=student).execute()
-        print('done')
-    except HttpError as e:
-        error = simplejson.loads(e.content).get('error')
-        if(error.get('code') == 409):
-            print('User {0} are already a member of this course.'.format(
-                student['userId']))
-        else:
-            raise
 
 
 #
 # enroll students to a classroom with no confirmation cannot execute
-# due to lack of authority for classroom API
-
-
-def enrollStudent2(courseId, studentId):
+# due to lack of authority(teacher / classroom domain != student domain).
+#
+def enrollUser(courseId, userId):
+    # test code: trying multiuser in 1 request.
+    #student = []
+    # for userId in userIds:
+    #    student.append({
+    #        'courseId': courseId,
+    #        'userId': userId,
+    #        'role': 'STUDENT'
+    #    })
+    # print(student)
     student = {
-        'userId': studentId
+        'userId': userId
     }
     try:
         student = serviceEnrollment.courses().students().create(
@@ -290,7 +293,7 @@ def enrollStudent2(courseId, studentId):
             print('User {0} are already a member of this course.'.format(
                 student.get('profile').get('name').get('fullName')))
         elif(error.get('code') == 403):
-            print('Permission Denied.')
+            print('...Permission Denied.')
         else:
             print(error.get('code'))
             raise
@@ -310,17 +313,22 @@ def listClassroom():
             # csv indexes
             writer.writerow(
                 ['courseName', 'courseSection', 'courseId', 'classCode', 'ownerId', 'teacherName', 'status'])
-            for course in tqdm(courses):
-                results = serviceEnrollment.courses().teachers().list(
-                    courseId=course.get('id')).execute()
-                teachers = results.get('teachers', [])
-                writer.writerow([course.get('name'), course.get('section'), course.get(
-                    'id'), course.get('enrollmentCode'), course.get('ownerId'), teachers[0]['profile']['name']['fullName'], course.get('courseState')])
-                # print(u'{0}, {1}, {2}, {3}, {4}, {5}'.format(course.get('name'), course.get('section'), course.get(
-                #    'id'), course.get('enrollmentCode'), course.get('ownerId'), teachers[0]['profile']['name']['fullName']))
-            # print(teachers[0]['profile']['emailAddress'])
-            # for teacher in teachers:
-            #    print('teacher: {}'.format(teacher['profile'])) # .get('name').get('fullName')
+            pool = Pool(processes = maxProcess)
+            multi_args = []
+            for course in courses:
+                multi_args.append([course])
+            results = pool.starmap(listClassroomMulti, multi_args)
+            writer.writerows(results)
+
+def listClassroomMulti(course):
+    results = serviceEnrollment.courses().teachers().list(courseId=course.get('id')).execute()
+    print(u'{}..'.format(course.get('id')))
+    teachers = results.get('teachers', [])
+    teacherName = ""
+    for teacher in teachers:
+        teacherName += '/' + str(teacher['profile']['name']['fullName'])
+    return([course.get('name'), course.get('section'), course.get('id'), course.get('enrollmentCode'), course.get('ownerId'), teacherName.lstrip('/'), course.get('courseState')])
+
 
 
 def infoClassroom(courseId):
@@ -330,16 +338,13 @@ def infoClassroom(courseId):
 
 # main()
 courseIdFile = "coursesID.csv"
+maxProcess = 10
 
 if __name__ == '__main__':
     global adminUser
     global options
     execMode, options = parseOptions()
     #print(execMode, options)
-    # evaluate specified key of dict object
-    # if 'enrollFile' in options:
-    #    print(options['enrollFile'])
-    # exit()
     # load config.ini
     inifile = configparser.ConfigParser()
     inifile.read('./config.ini', 'UTF-8')
@@ -356,7 +361,9 @@ if __name__ == '__main__':
         target = classSubjects
     elif execMode == 'remove':
         for courseId in options['courses']:
+            print('removing.. {}'.format(courseId), end="")
             deleteClassroom(courseId)
+            print('done')
         exit()
     elif execMode == 'lists':
         listClassroom()
@@ -379,7 +386,7 @@ if __name__ == '__main__':
             print('Course  Name:{}'.format(classSubject))
             print('Subject Name:{}'.format(classSection))
             print('Lecturer    :{}'.format(classTeacher))
-            if not option['dry-run']:
+            if not options['dry-run']:
                 # create Classroom ownered by class teacher.
                 courseId, enrollCode = createClassroom(
                     classSubject, classSections[classCode], classTeacher)
@@ -389,19 +396,29 @@ if __name__ == '__main__':
         # addAdminUser(courseId)
         #
         if execMode == 'enroll' or execMode == 'default':
-            print('enrolling..')
             # add adminUser while students are added to a course
-            if classCode in enrollStudents:
-                if classTeacher != adminUser and not options['dry-run']:
+            if classCode in enrollUsers:
+                if classTeacher != adminUser and not options['dry-run'] and 'foreignDomain' in options:
                     addAdminUser(courseId)
-                # print(enrollStudents[classCode])
-                for studentId in enrollStudents[classCode]:
-                    studentEmail = studentEmails[studentId]
-                    print(classCode, studentId, end="")
-                    if not option['dry-run']:
-                        enrollStudent(courseId, studentEmail)
-                    # print(studentEmail, end="")
-                if classTeacher != adminUser and not options['dry-run']:
+                # print(enrollUsers[classCode])
+                for userId in enrollUsers[classCode]:
+                    if not 'teacherRole' in options:
+                        studentEmail = studentEmails[userId]
+                    print(classCode, userId, end="")
+                    if not options['dry-run']:
+                        print(' is enrolling.. ', end="")
+                        if 'teacherRole' in options:
+                            if 'foreignDomain' in options:
+                                inviteUser(courseId, userId, 'TEACHER' )
+                            else:
+                                enrollTeacer(courseId, userId)
+                        else:
+                            if 'foreignDomain' in options:
+                                inviteUser(courseId, studentEmail, 'STUDENT')
+                            else:
+                                enrollUser(courseId, studentEmail)
+                            # print(studentEmail, end="")
+                if classTeacher != adminUser and not options['dry-run'] and 'foreignDomain' in options:
                     deleteAdminUser(courseId)
     if not options['dry-run']:
         file.close()
